rename solaris::sol module to solaris::convert since its a more fitting name
