+++ solaris/src/convert.rs
